Enabling travis build

+++ deploy.sh
@@ -7,23 +7,33 @@ shopt -s extglob
 
 if [ $TRAVIS_PULL_REQUEST = false ]
 then
-    echo "targetBranch = master"
-    git config --global user.email $GIT_USER_EMAIL
-    git config --global user.name $GIT_USER_NAME
+    echo "current branch is =  $TRAVIS_BRANCH"
+    if [ $TRAVIS_BRANCH = "develop" ]
+    then
+        echo "targetBranch = master"
+        git config --global user.email $GIT_USER_EMAIL
+        git config --global user.name $GIT_USER_NAME
 
-    echo "Starting deployment ... "
+        echo "Starting deployment ... "
 
-    git clone https://$USER_CRED@github.com/house-lending-apps/HouseLendingDeployable.git $targetBranch
-    git checkout $targetBranch
+        cd ..
+        echo "Current directory = $PWD"
+        git clone https://$USER_CRED@github.com/house-lending-apps/HouseLendingDeployable.git $targetBranch
+        git checkout $targetBranch
 
-    cd HouseLendingDeployable
+        echo "ls -l"
+        cd HouseLendingDeployable
+        echo "Current directory = $PWD"
 
-    cp -r ../$REPOSITORY_NAME/dist/. .
-    echo "Pushing to $targetBranch"
+        cp -r ../$REPOSITORY_NAME/dist/. .
+        echo "Pushing to $targetBranch"
 
-    git add -f -all
-    git commit -a -m 'bumped the version [ci skip]'
-    git push -f --set-upstream origin $targetBranch
+        git add -f --all .
+        git commit -a -m 'bumped the version [ci skip]'
+        git push -f --set-upstream origin $targetBranch
+    else
+        echo 'No need to deploy.'
+    fi
 
 else
         echo 'No need to deploy.'
